requirements.txt
# Requirements updated: 2025-05-10 12:19:30
PyYAML==6.0.2
SQLAlchemy==2.0.40
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
colormath==3.0.0
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.6.0
greenlet==3.2.1
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx-sse==0.4.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langchain==0.3.25
langsmith==0.3.42
marshmallow==3.26.1
mcp==1.7.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
nvidia-cusparselt-cu12==0.6.3
openai-agents==0.0.14
openai==1.77.0
orjson==3.10.18
packaging==24.2
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
propcache==0.3.1
pydantic-settings==2.9.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
redis==6.0.0
regex==2024.11.6
requests-toolbelt==1.0.0
requests==2.32.3
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
triton==3.3.0
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
